+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Upcoming Version
 
+- Bugfix: Fix command line handling when using the `git remote` command.
+
 ## v1.0.25
 
 - Handle `git remote add origin` by cloning the remote if it was not mirrored